requirements.txt
Django==2.2
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
requests==2.21.0
waitress==1.2.1